requirements.txt
torch>=2.2.0
torchvision>=0.17.0
pytorch-lightning>=2.3.0
webdataset>=0.2.86
boto3>=1.34.0
wandb>=0.16.0
timm>=0.9.16
omegaconf>=2.3.0
Pillow>=10.0.0
numpy>=1.24.0
braceexpand>=0.1.7
